DDCNLS-5055 | refactored transformer code
