+++ app/models/AtsMiddleTierData.scala
@@ -26,4 +26,33 @@ case class AtsMiddleTierData(taxYear: Int, utr: Option[String], income_tax: Opti
 
 object AtsMiddleTierData {
   implicit val formats:Format[AtsMiddleTierData] = Json.format[AtsMiddleTierData]
+
+  def make(
+    taxYear: Int,
+    utr: String,
+    incomeTax: DataHolder,
+    summary: DataHolder,
+    income: DataHolder,
+    allowance: DataHolder,
+    capitalGains: DataHolder,
+    govSpending: GovernmentSpendingOutputWrapper,
+    taxPayer: AtsMiddleTierTaxpayerData
+  ): AtsMiddleTierData =
+    AtsMiddleTierData(
+      taxYear,
+      Some(utr),
+      Some(incomeTax),
+      Some(summary),
+      Some(income),
+      Some(allowance),
+      Some(capitalGains),
+      Some(govSpending),
+      Some(taxPayer),
+      None
+    )
+
+  def error(taxYear: Int, message: String): AtsMiddleTierData =
+    AtsMiddleTierData(taxYear, None, None, None, None, None, None, None, None, Option(AtsError(message)))
+
+  def noAtsResult(taxYear: Int): AtsMiddleTierData = error(taxYear, "NoAtsError")
 }
