+++ app/transformers/GovSpendingDataTransformer.scala
@@ -19,7 +19,7 @@ package transformers
 import models._
 import services.GovSpendService
 
-case class GovSpendingDataTransformer(totalTaxAmount: Amount, taxYear: Int){
+case class GovSpendingDataTransformer(totalTaxAmount: Amount, taxYear: Int) {
 
   lazy val govSpendReferenceDTO = createGovSpendingReferenceDTO
   
@@ -27,7 +27,7 @@ case class GovSpendingDataTransformer(totalTaxAmount: Amount, taxYear: Int){
 
   def createSpendDataItem(spendCategory: String, spendPercentage: BigDecimal, amount: Amount): SpendData = {
     val monetaryBD = getMonetaryAmount(spendPercentage, amount)
-    val monetaryAmount = new Amount(monetaryBD)
+    val monetaryAmount = Amount.gbp(monetaryBD)
     SpendData(monetaryAmount, spendPercentage)
   }
 
@@ -35,8 +35,6 @@ case class GovSpendingDataTransformer(totalTaxAmount: Amount, taxYear: Int){
     case (key, value) =>
       (key, createSpendDataItem(key, value, totalTaxAmount))
   }
-
-  private def getPercentage(key: String) = GovSpendService.govSpending(taxYear).get(key).get
 
   private def getMonetaryAmount(percentage: BigDecimal, amount: Amount) = {
     ((percentage / 100) * amount.amount).setScale(2, BigDecimal.RoundingMode.HALF_UP)
